Revert "Swapped iFFBFFull over to no observation error."
This reverts commit 2c0a1fa6ac635a595ca14c95fca098a19ba92a47.

This was done since for no observation error we need to
initialise correctly, which doesn't work because
we've added P into the mix. If we removed this then it will
work, or we need to initialise to be consistent with
P and well as the end nodes. I've reverted rather than removed
the original commit so that these changes are in the repo for
posterity, in case they want to be resurrected in the future.